+++ Tests/iLogYouTests/TestFileGeneration.swift
@@ -1,0 +1,24 @@
+//
+//  CITest.swift
+//  iLogYouTests
+//
+//  Created by pierre piron on 05/02/2019.
+//
+
+import XCTest
+@testable import iLogYou
+
+class TestFileGeneration: XCTestCase {
+    
+    func testCreateValidAppDelegateFile() {
+        let appDelegate = BasicFiles.appDelegate
+        XCTAssertTrue(appDelegate.contains("class AppDelegate"))
+        XCTAssertTrue(appDelegate.contains("func application("))
+        XCTAssertTrue(appDelegate.contains("func applicationWillResignActive"))
+        XCTAssertTrue(appDelegate.contains("func applicationDidEnterBackground("))
+        XCTAssertTrue(appDelegate.contains("func applicationWillEnterForeground("))
+        XCTAssertTrue(appDelegate.contains("func applicationDidBecomeActive("))
+        XCTAssertTrue(appDelegate.contains("func applicationWillTerminate("))
+    }
+    
+}
